Move hooks from `virtool_workflow_runtime` to `virtool_workflow`.
Keep all the hooks available in one module.
